RF_02.007.04_POM Verify “Yes” color deleting Pipeline using sidebar (#1236)
