+++ src/test/java/school/redrover/model/DeleteDialog.java
@@ -1,11 +1,11 @@
 package school.redrover.model;
 
+import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import school.redrover.model.base.BasePage;
-import school.redrover.runner.TestUtils;
 
 public class DeleteDialog extends BasePage {
     public DeleteDialog(WebDriver driver) {
@@ -19,5 +19,13 @@ public class DeleteDialog extends BasePage {
     public <T> T clickYes(T page) {
         getWait10().until(ExpectedConditions.elementToBeClickable(yesButton)).click();
         return page;
+    }
+
+    public String getYesButtonColorDeletingViaSidebar() {
+        JavascriptExecutor js = (JavascriptExecutor) getDriver();
+
+        return (String) js.executeScript(
+                "return window.getComputedStyle(arguments[0]).getPropertyValue('--color');",
+                yesButton);
     }
 }
