+++ src/test/java/school/redrover/model/base/BasePage.java
@@ -5,7 +5,6 @@ import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
-import school.redrover.model.AppearancePage;
 import school.redrover.model.HomePage;
 
 public abstract class BasePage extends BaseModel {
@@ -25,18 +24,6 @@ public abstract class BasePage extends BaseModel {
 
         ((JavascriptExecutor) getDriver()).executeScript("arguments[0].dispatchEvent(new Event('mouseenter'));", chevron);
         ((JavascriptExecutor) getDriver()).executeScript("arguments[0].dispatchEvent(new Event('click'));", chevron);
-    }
-
-    public AppearancePage resetJenkinsTheme() {
-        getDriver().findElement(By.cssSelector("[href='/manage']")).click();
-        getDriver().findElement(By.cssSelector("[href='appearance']")).click();
-
-        WebElement defaultThemeButton = getDriver().findElement(By.cssSelector("[for='radio-block-2']"));
-        if (!defaultThemeButton.isSelected()) {
-            defaultThemeButton.click();
-            getDriver().findElement(By.name("Apply")).click();
-        }
-        return new AppearancePage(getDriver());
     }
 
     public void openHeaderUsernameDropdown() {
